refactor(daemon): drop Cloudflare Access JWT middleware
The CF Access JWT path added significant complexity for a feature most
deployments never used. Bearer token (OD_ACCESS_TOKEN) and trusted
reverse proxy (OD_TRUSTED_PROXY) cover the same identity surface with
operator-trust semantics instead of per-request JWT validation.

- Replace apps/daemon/src/cf-access-middleware.ts with a minimal
  apps/daemon/src/auth/auth-mode.ts that resolves one of
  'none' | 'trusted-proxy' | 'access-token' from env.
- Remove CF imports, startup wiring, and middleware mount from
  apps/daemon/src/server.ts.
- OD_BEHIND_PROXY=cloudflare, OD_CF_ACCESS_TEAM_DOMAIN,
  OD_CF_ACCESS_AUD, and OD_CF_ACCESS_UNSAFE_DOMAIN are no longer
  recognized auth signals.

Cloudflare can still gate Open Design as a generic trusted proxy via
OD_TRUSTED_PROXY=cloudflare; identity enforcement stays at the proxy.
